docs(core): update XML doc comments across public API

+++ src/TraceQuery.Core/Model/TraceEntry.cs
@@ -10,12 +10,12 @@ namespace TraceQuery.Core.Model;
 /// </summary>
 public class TraceEntry
 {
-    /// <summary>Constructor for TraceEntry class.</summary>
+    /// <summary>Sets the attributes of a TraceEntry object.</summary>
     /// <param name="timestamp">Date and time of event with UTC offset.</param>
     /// <param name="severity">Severity classification of the event.</param>
     /// <param name="component">ID of component associated with the event.</param>
     /// <param name="message">Event description.</param>
-    public TraceEntry(DateTimeOffset timestamp, Severity severity, Byte component, String? message)
+    public TraceEntry(DateTimeOffset timestamp, Severity severity, int component, String? message)
     {
         Timestamp = timestamp;
         Severity = severity;
@@ -30,12 +30,12 @@ public class TraceEntry
     public Severity Severity { get; }
 
     /// <summary>ID of component associated with the event.
-    /// Justification of Byte type: suspected memory overhead of storing many strings.
+    /// Justification of Int type: suspected memory overhead of storing many strings.
     /// Possible computational overhead of comparing TraceEntry components on analysis.
     /// TODO: component->id (ingestion), id->component (reporting)
     /// note: Complexity introduced without profiling. Monitor debt carefully.
     /// </summary>
-    public Byte Component { get; }
+    public int Component { get; }
 
     /// <summary>Event description.</summary>
     public String? Message { get; }
